Improve the analysis frame rate selection.
- Don't assume a video frame rate of 30fps. This is not always true!
- Select a valid initial analysis frame rate close to 10fps if the video
frame rate is unequal 30fps.

+++ src/nz/ac/auckland/lablet/camera/CalibrationVideoTimeData.java
@@ -10,14 +10,14 @@ package nz.ac.auckland.lablet.camera;
 
 public class CalibrationVideoTimeData implements ITimeData {
     protected int numberOfFrames;
-    final protected int videoDuration;
+    final protected long videoDuration;
 
     protected float analysisFrameRate;
     // milli seconds
     protected float analysisVideoStart;
     protected float analysisVideoEnd;
 
-    public CalibrationVideoTimeData(int videoDuration) {
+    public CalibrationVideoTimeData(long videoDuration) {
         this.videoDuration = videoDuration;
 
         setAnalysisVideoStart(0);
